requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
websockets>=12.0
motor>=3.3.2
pymongo>=4.6.0
httpx>=0.25.2
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
skyfield>=1.47
poliastro>=0.17.0
astropy>=5.3.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
celery>=5.3.0
redis>=5.0.0
plotly>=5.17.0
aiofiles>=23.2.0